+++ src/kudu/security/tls_socket.h
@@ -40,7 +40,7 @@ class TlsSocket : public Socket {
 
   Status Writev(const struct ::iovec *iov,
                 int iov_len,
-                int32_t *nwritten) override WARN_UNUSED_RESULT;
+                int64_t *nwritten) override WARN_UNUSED_RESULT;
 
   Status Recv(uint8_t *buf, int32_t amt, int32_t *nread) override WARN_UNUSED_RESULT;
